chp_api/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
backports.csv==1.0.7
backports.tarfile==1.1.1
beautifulsoup4==4.12.3
billiard==4.2.0
bmt==1.4.0
cattrs==23.2.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cheroot==10.0.1
CherryPy==18.9.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
curies==0.7.9
Deprecated==1.2.14
deprecation==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-hosts==6.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
feedparser==6.0.11
flower==2.0.1
future==1.0.0
hbreader==0.9.1
humanize==4.9.0
idna==3.7
importlib-metadata==7.0.0
inflect==7.2.0
iniconfig==2.0.0
isodate==0.6.1
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.0
joblib==1.4.0
json-flattener==0.1.9
jsonasobj2==1.0.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kombu==5.3.7
linkml-runtime==1.7.5
lxml==5.2.1
more-itertools==10.2.0
mysqlclient==2.2.4
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-django==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
packaging==24.0
pandas==2.2.2
Pattern==3.6
pdfminer.six==20231228
platformdirs==4.2.1
pluggy==1.5.0
portend==3.2.0
prefixcommons==0.1.12
prefixmaps==0.2.4
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==1.10.12
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.1.1
pytest-logging==2015.11.4
python-dateutil==2.9.0.post0
python-docx==1.1.0
PyTrie==0.4.0
pytz==2024.1
PyYAML==6.0.1
rdflib==7.0.0
reasoner-pydantic==5.0.2
redis==5.0.4
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-cache==1.2.0
rpds-py==0.18.0
scipy==1.13.0
sgmllib3k==1.0.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stringcase==1.2.0
tempora==5.5.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
uWSGI==2.0.25.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
zc.lockfile==3.0.post1
zipp==3.18.1
